% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsurvey.R
\name{as.dtsurvey}
\alias{as.dtsurvey}
\title{Convert a survey::svydesign object into a dtsurvey}
\usage{
as.dtsurvey(sur)
}
\arguments{
\item{sur}{svydesign object
This function converts a survey.design2 object into a dtsurvey object.
Finite population corrections (fpcs) are not implemented in dtsurvey yet}
}
\description{
Convert a survey::svydesign object into a dtsurvey
}
