% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcdtsurvey.R
\name{calc.dtsurvey}
\alias{calc.dtsurvey}
\title{Compute CHI type estimates from a dtsurvey object}
\usage{
\method{calc}{dtsurvey}(
  ph.data,
  what,
  where,
  by = NULL,
  metrics = c("mean", "numerator", "denominator"),
  per = NULL,
  win = NULL,
  time_var = NULL,
  proportion = FALSE,
  fancy_time = TRUE,
  ci = 0.95,
  verbose = FALSE
)
}
\description{
@export
}
