% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{calc_mean_dtrepsurvey}
\alias{calc_mean_dtrepsurvey}
\title{An internal function to calculate the mean (and optionally, SE and CI) from a dtrepsurvey object
https://github.com/cran/survey/blob/a0f53f8931f4e304af3c758b2ff9a56b0a2a49bd/R/surveyrep.R}
\usage{
calc_mean_dtrepsurvey(x, sv, scaledata, cw, mse, var = T)
}
\description{
An internal function to calculate the mean (and optionally, SE and CI) from a dtrepsurvey object
https://github.com/cran/survey/blob/a0f53f8931f4e304af3c758b2ff9a56b0a2a49bd/R/surveyrep.R
}
