% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{calculate_error}
\alias{calculate_error}
\title{Calculate error statistics for survey means}
\usage{
calculate_error(ret, ci_method, level, df, var_type, lids)
}
\arguments{
\item{ret}{list. contains the results from calc_mean/total_...}

\item{ci_method}{character. Type of ci (if any, to calculate)}

\item{level}{numeric. 0 - 1, confidence level to return for ci}

\item{df}{numeric. degrees of freedom}

\item{var_type}{character. type of variability/variance to return}

\item{lids}{numeric. Length of ids (e.g. number of rows of data to consider)}
}
\description{
Calculate error statistics for survey means
}
