% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{ci_score}
\alias{ci_score}
\title{CI of a proportion via score method}
\usage{
ci_score(success, N = sum(success), level = 0.95)
}
\arguments{
\item{success}{numeric. Vector of counts}

\item{N}{numeric. Single number of vector of N such that success/N = proportions}

\item{level}{numeric [0-1]. Confidence level.}
}
\description{
CI of a proportion via score method
}
