% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{ci_standard}
\alias{ci_standard}
\title{What you'd get from survey mean}
\usage{
ci_standard(x, se, level, df, survey = T)
}
\arguments{
\item{x}{numeric. THe mean/total result}

\item{se}{numeric. The standard error}

\item{level}{numeric. 0 - 1 reflecting the confidence level}

\item{df}{numeric. degrees of freedom}

\item{survey}{logical. Specifies whether the data comes from a survey.}
}
\description{
What you'd get from survey mean
}
