% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{ci_xlogit}
\alias{ci_xlogit}
\title{Refer to xlogit option of svyciprop}
\usage{
ci_xlogit(x, vcov, level, df, st)
}
\arguments{
\item{x}{numeric. THe mean/total result}

\item{vcov}{matrix The variance/covariance matrix}

\item{level}{numeric. 0 - 1 reflecting the confidence level}

\item{df}{numeric. degrees of freedom}

\item{st}{character. Survey type, one of svyrepdt or svydt}
}
\description{
Refer to xlogit option of svyciprop
}
