% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcdtsurvey.R
\name{compute}
\alias{compute}
\title{A function to compute a metric as part of calc.dtsurvey}
\usage{
compute(
  DT,
  x,
  metrics,
  ci_method = "mean",
  level = 0.95,
  time_var,
  time_format
)
}
\arguments{
\item{DT}{data.table to compute a metric from}

\item{metrics}{character vector. Vector of metrics to compute}

\item{ci_method}{One of 'mean', TODO: Fill in the rest determining how the CI/se should be generated}

\item{level}{numeric between 0 and 1. Lelvel of confidence}
}
\description{
A function to compute a metric as part of calc.dtsurvey
}
