% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtadmin.R
\name{dtadmin}
\alias{dtadmin}
\title{Convert a data.table into a dtadmin, dtsurvey object to use some additional functionality}
\usage{
dtadmin(DT, copy = TRUE)
}
\arguments{
\item{DT}{data.table}

\item{copy}{logical. Determines whether a copy of DT is taken before setting new attributes.}
}
\description{
Convert a data.table into a dtadmin, dtsurvey object to use some additional functionality
}
\details{
dtadmin by default copies the dataset (not be reference)
}
