% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrepsurvey.R
\name{dtrepsurvey}
\alias{dtrepsurvey}
\alias{dtrepsurvey.data.frame}
\alias{dtrepsurvey.svyrep.design}
\title{A function to create a dtsvyrep from a data.frame or svyrep.design type object}
\usage{
dtrepsurvey(svyrep, ...)

\method{dtrepsurvey}{data.frame}(svyrep, ...)

\method{dtrepsurvey}{svyrep.design}(svyrep, ...)
}
\arguments{
\item{svyrep}{svyrep.design or an object that inherits data.frame.}

\item{...}{options passed to survey::svrepdesign when svyrep is a data.frame object.
Ignored for svyrep which already contains the necessary information for transformation}
}
\description{
A function to create a dtsvyrep from a data.frame or svyrep.design type object
}
