% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsurvey.R
\name{dtsurvey}
\alias{dtsurvey}
\title{Construct a dtsurvey object}
\usage{
dtsurvey(DT, psu = NULL, strata = NULL, weight = NULL, nest = TRUE)
}
\arguments{
\item{DT}{data.table. A data.table containing the survey results}

\item{psu}{character. Name of the variable(s) containing the psu information. NULL indicates no PSUs (e.g. each observation gets its own)}

\item{strata}{character. Name of the variable(s) containing the strata information. Null indicates no strata (e.g. all obs belong to the same one)}

\item{weight}{character. Name of the variables containing the pweight information. If NULL, then weights default to 1 (and what even is the point then)}

\item{nest}{logical. If true, will re-nest PSUs within strata.
If false, PSUs will be accepted as is and will throw an error when PSUs belong to multiple strata}
}
\description{
Construct a dtsurvey object
}
