% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{[.dtsurvey}
\title{Custom extract function (or subset or whatever) for dtsurveys. A loose wrapper over \verb{[.data.table}}
\usage{
\method{[}{dtsurvey}(x, i, j, by, ..., drop = NULL)
}
\arguments{
\item{x}{a dtsurvey object}

\item{i}{i in the data.table format}

\item{j}{j in the data.table format}

\item{by}{by in the data.table format}

\item{...}{extra options passed to data table}

\item{drop}{used to make dplyr things work nice. should generally be ignored}
}
\description{
Custom extract function (or subset or whatever) for dtsurveys. A loose wrapper over \verb{[.data.table}
}
