% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_type.R
\name{get_survey_type}
\alias{get_survey_type}
\title{Get survey type}
\usage{
get_survey_type()
}
\description{
Get survey type
}
\details{
This function is only designed to be called by other functions within J of a data.table.
It extracts the stype attribute data.table from the calling data.table to be used for calculating the survey metrics
}
