% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.dtsurvey}
\alias{merge.dtsurvey}
\title{Merge information on to a dtsurvey}
\usage{
\method{merge}{dtsurvey}(x, y, ...)
}
\arguments{
\item{x}{data.table}

\item{y}{data.frame or similar object}

\item{...}{options passed to merge.data.table}
}
\description{
Merge information on to a dtsurvey
}
