% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{prep_survey_data}
\alias{prep_survey_data}
\title{Prepare data for survey calculations}
\usage{
prep_survey_data(var_type, ci_method, ids, use_df, na.rm, x)
}
\arguments{
\item{var_type}{character.}

\item{ci_method}{character.}

\item{ids}{vector.}

\item{use_df}{logical.}

\item{na.rm}{logical.}

\item{x}{vector}
}
\description{
Prepare data for survey calculations
}
\details{
This function is merely a helper for \code{smean} and \code{stotal}.
}
