% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{repdes_var}
\alias{repdes_var}
\title{Calculate the variance/covariance matrix for a survey statistic (replicate design)}
\usage{
repdes_var(x, type = "mean", ids, sv, svyrep_attributes)
}
\arguments{
\item{x}{vector of data}

\item{type}{character. One of 'mean' or 'total' depending on the scale of statistic}

\item{ids}{numeric. vector of indices to operate on}

\item{sv}{data.table. Survey vars data.table}

\item{svyrep_attributes}{list. List of attributes relevant to replicate design surveys}
}
\description{
Calculate the variance/covariance matrix for a survey statistic (replicate design)
}
