% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{smean}
\alias{smean}
\alias{smean.default}
\alias{smean.character}
\alias{stotal.character}
\title{Calculate the survey mean}
\usage{
smean(x, ...)

\method{smean}{default}(
  x,
  na.rm = T,
  var_type = "none",
  ci_method = "mean",
  level = 0.95,
  use_df = T,
  ids,
  sv,
  st,
  ...
)

\method{smean}{character}(x, ...)

\method{stotal}{character}(x, ...)
}
\arguments{
\item{x}{vector. Vector of values to compute a weighted mean over}

\item{...}{other arguments. Currently unused.}

\item{na.rm}{logical. Determines whether NAs are excluded from the analysis}

\item{var_type}{character. Report variability as one or more of: standard error ("se", default), confidence interval ("ci") and variance-covariance matrix ('vcov')}

\item{ci_method}{character. Determines how the ci (if requested via \code{var_type}) should be calculated
Options beside "mean" are only relevant for proportion (e.g. logical or values).
When the method is 'mean', the results should match \code{survey::svymean} while other options match \code{survey::svyciprop}}

\item{level}{numeric. A value in the range of (0, 1) denoting what level of confidence the CI should be}

\item{use_df}{logical. Should the estimated degrees of freedom by used to calculate CIs? Default is TRUE. FALSE implies df = Inf.
\code{confint(survey::svymean())} uses Inf as a default while \code{confint(survey::svyciprop)} uses \code{degf(design)}}

\item{ids}{numeric. indices which are being computed. Can be generally omitted and will be added to the call via \verb{[.dtsurvey}}

\item{sv}{data.table. Data.table of psu, strata, weight and the like to properly do survey statistics.}

\item{st}{character. type of survey dataset being analyzed. Can be generally omitted and will be added to the call via \verb{[.dtsurvey}}
}
\value{
a vector or a list (the latter likely converted into a data.table) containing the results
}
\description{
Calculate the survey mean
}
\details{
If var_type is "none", a vector is returned. For factors, the names of the vector correspond to the levels. When var_type is not "none" a named list is returned (and then likely converted into a data.table).
The list (which should be named) is ordered in the following manner (by slot): result, se, lower, upper, levels.
}
