% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_assignment.R
\name{stop_assignment}
\alias{stop_assignment}
\title{Stop a call in a dtsurvey if its part of an assignment}
\usage{
stop_assignment()
}
\description{
Stop a call in a dtsurvey if its part of an assignment
}
