% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{sur_ci}
\alias{sur_ci}
\title{Calculate confidence interval}
\usage{
sur_ci(
  a,
  b = "sur_mean",
  ab_type = "raw",
  ci_part = "both",
  ci_method = "mean",
  level = 0.95,
  use_df = T,
  denom,
  na.rm = T,
  sv,
  ids,
  st
)
}
\arguments{
\item{a}{Either the result of sur_mean or sur_total OR a vector of data to compute a function over}

\item{b}{Either a vcov matrix from sur_var OR one of 'sur_total' or 'sur_mean'}

\item{ab_type}{character. One of "agg" or "raw". If agg, \code{a} is the result and \code{b} is the corresponding vcov matrix. If 'raw', \code{a} is the vector to compute \code{b} on (e.g. \code{b}(\code{a}))}

\item{ci_part}{character. One of 'lower', 'upper', 'both'. Determines what part of the ci should be returned}

\item{ci_method}{character. Name of the method by which the ci should be calculated. See details for more information.}

\item{level}{numeric. 0 - 1, confidence level to return for ci}

\item{use_df}{logical}

\item{denom}{numeric. Only useful/used when ci_method == 'unweighted_binary'. Denominator (or N) for prop.test}

\item{na.rm}{logical. Passed to \code{b} when \code{ab_type} = 'raw'}

\item{sv}{data.table. Survey vars data.table}

\item{ids}{numeric. vector of indices to operate on}

\item{st}{character. survey type}
}
\description{
Calculate confidence interval
}
