% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_mean.R
\name{sur_mean}
\alias{sur_mean}
\alias{sur_mean.default}
\alias{sur_mean.factor}
\alias{sur_mean.character}
\title{Calculate the survey weighted mean of an object}
\usage{
sur_mean(x, ...)

\method{sur_mean}{default}(x, na.rm = T, sv, ids, st)

\method{sur_mean}{factor}(x, na.rm = T, add.levels = FALSE, sv, ids, st)

\method{sur_mean}{character}(x, ...)
}
\arguments{
\item{x}{vector. Variable to compute a mean}

\item{na.rm}{logical. Determines whether NA values will be omitted from the analysis}

\item{sv}{data.table. Survey vars data.table}

\item{ids}{numeric. vector of indices to operate on}

\item{st}{character. survey type}

\item{add_levels}{logical. Determines whether the result should return a two item list (the results, level names)}
}
\description{
Calculate the survey weighted mean of an object
}
