% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_total.R
\name{sur_total}
\alias{sur_total}
\alias{sur_total.default}
\alias{sur_total.factor}
\alias{sur_total.character}
\title{Calculate the survey weighted total of an object}
\usage{
sur_total(x, ...)

\method{sur_total}{default}(x, na.rm = T, as_list = FALSE, sv, ids, st)

\method{sur_total}{factor}(x, na.rm = T, as_list = FALSE, sv, ids, st)

\method{sur_total}{character}(x, ...)
}
\arguments{
\item{x}{vector. Variable to compute a total}

\item{na.rm}{logical. Determines whether NA values will be omitted from the analysis}

\item{as_list}{logical. Determines whether the result should returned as a list-- mostly for factors}

\item{sv}{data.table. Survey vars data.table}

\item{ids}{numeric. vector of indices to operate on}

\item{st}{character. survey type}
}
\description{
Calculate the survey weighted total of an object
}
