% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{sur_var}
\alias{sur_var}
\title{Calculate the variance/covariance matrix for a survey statistic}
\usage{
sur_var(
  x,
  na.rm = T,
  type = "mean",
  as_list = TRUE,
  svyrep_attributes,
  sv,
  ids,
  st
)
}
\arguments{
\item{x}{vector of data}

\item{na.rm}{logical. If TRUE, omit NAs}

\item{type}{character. One of 'mean' or 'total' depending on the scale of statistic}

\item{as_list}{logical. Return values as list}

\item{svyrep_attributes}{list. List of attributes relevant to replicate design surveys}

\item{sv}{data.table. Survey vars data.table}

\item{ids}{numeric. vector of indices to operate on}

\item{st}{character. survey type}
}
\description{
Calculate the variance/covariance matrix for a survey statistic
}
