% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{surdes_var}
\alias{surdes_var}
\title{Calculate the variance/covariance matrix for a survey statistic (normal survey)}
\usage{
surdes_var(x, type = "mean", ids, sv)
}
\arguments{
\item{x}{matrix of data}

\item{type}{character. One of 'mean' or 'total' depending on the scale of statistic}

\item{ids}{numeric. vector of indices to operate on}

\item{sv}{data.table. Survey vars data.table}
}
\description{
Calculate the variance/covariance matrix for a survey statistic (normal survey)
}
