% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{calc_mean_dtrepsurvey}
\alias{calc_mean_dtrepsurvey}
\title{An internal function to calculate the mean (and optionally, SE and CI) from a dtrepsurvey object}
\usage{
calc_mean_dtrepsurvey(x, sv, ids, scaledata, cw, mse, selfrep = NULL, var = T)
}
\arguments{
\item{x}{matrix. Results from the mean call}

\item{sv}{data.table survey variables data.table}

\item{ids}{logical. vector determining which rows are being calculated on.}

\item{scaledata}{list the different scale bits from svrepdesigns}

\item{cw}{logical. denotes combined weights}

\item{mse}{logical. passed to survey::svrVar}

\item{selfrep}{logical. Used for some logic taken from survey:::svymean.svrepdesign}

\item{var}{logical. Should the vcov matrix be returned?}
}
\description{
An internal function to calculate the mean (and optionally, SE and CI) from a dtrepsurvey object
}
