% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{delist_factor}
\alias{delist_factor}
\title{Unlist by a single level, used for survey factors results}
\usage{
delist_factor(x)
}
\description{
Unlist by a single level, used for survey factors results
}
