% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrepsurvey.R
\name{dtrepsurvey}
\alias{dtrepsurvey}
\title{A function to convert a `svyrep.design` into a dtsvyrep}
\usage{
dtrepsurvey(svyrep)
}
\arguments{
\item{svyrep}{svyrep.design}
}
\description{
A function to convert a `svyrep.design` into a dtsvyrep
}
