% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_vars.R
\name{get_survey_vars}
\alias{get_survey_vars}
\title{Get survey vars}
\usage{
get_survey_vars()
}
\description{
Get survey vars
}
\details{
This function is only designed to be called by other functions within J of a data.table.
It extracts the sdes attribute data.table from the calling data.table to be used for calculating the survey metrics
}
