% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{smean}
\alias{smean}
\alias{smean.numeric}
\alias{smean.default}
\alias{smean.factor}
\title{Calculate the survey mean}
\usage{
smean(x, ...)

\method{smean}{numeric}(x, na.rm = T, ids)

\method{smean}{default}(x, na.rm = T, ids)

\method{smean}{factor}(x, na.rm = T, ids)
}
\arguments{
\item{x}{vector Vector of values to compute a weighted mean over}

\item{na.rm}{logical. Passed to `weighted.mean`}

\item{ids}{numeric. list of indices which are being computed}
}
\description{
Calculate the survey mean
}
