% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{smean}
\alias{smean}
\alias{smean.default}
\alias{smean.character}
\title{Calculate the survey mean}
\usage{
smean(x, ...)

\method{smean}{default}(
  x,
  ids,
  na.rm = T,
  var_type = "none",
  ci_method = "mean",
  level = 0.95,
  df = Inf
)

\method{smean}{character}(x, ...)
}
\arguments{
\item{x}{vector. Vector of values to compute a weighted mean over}

\item{ids}{numeric. list of indices which are being computed}

\item{na.rm}{logical. Passed to `weighted.mean`}

\item{var_type}{character. Report variability as one or more of: standard error ("se", default) and confidence interval ("ci")}

\item{ci_method}{character. Determines how the ci (if requested via \code{var_type}) should be calculated
Options beside "mean" are only relevant for proportion (e.g. logical or values )}

\item{level}{numeric. A value in the range of (0, 1) denoting what level of confidence the CI should be}

\item{df}{numeric. Degrees of freedom}
}
\description{
Calculate the survey mean
}
