% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_calcs.R
\name{stotal}
\alias{stotal}
\alias{stotal.default}
\title{Calculate the survey total}
\usage{
stotal(x, ...)

\method{stotal}{default}(x, ids, na.rm = T, var_type = "none", level = 0.95, use_df = FALSE, ...)
}
\arguments{
\item{x}{vector. Vector of values to compute a weighted total over}

\item{...}{other arguments. Currently unused.}

\item{ids}{numeric. list of indices which are being computed.}

\item{na.rm}{logical. Determines whether NAs are removed from calculations}

\item{var_type}{character. Report variability as one or more of: standard error ("se", default) and confidence interval ("ci")}

\item{level}{numeric. A value in the range of (0, 1) denoting what level of confidence the CI should be}

\item{use_df}{logical. Should the estimated degrees of freedom by used to calculate CIs? Default is FALSE (different from \code{smean}. FALSE implies df = Inf.}
}
\value{
a list. Entry 1 is the result of the calculation (e.g. the mean value). Other item entries are (optionally) the se, lower, and upper.
}
\description{
Calculate the survey total
}
\details{
When x is a factor, results are returned in order of \code{levels(x)}.
         When \code{stotal} is called without assignment (e.g. \code{:=}), the result value will translate from a list to a data.table, even if only the mean is returned.
}
