% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_mean.R
\name{sur_mean}
\alias{sur_mean}
\alias{sur_mean.default}
\alias{sur_mean.factor}
\alias{sur_mean.character}
\alias{surdes_mean}
\alias{repdes_mean}
\alias{admin_mean}
\title{Calculate the survey weighted mean of an object}
\usage{
sur_mean(x, ...)

\method{sur_mean}{default}(x, na.rm = T, as_list = FALSE, sv, ids, st, ...)

\method{sur_mean}{factor}(x, na.rm = T, as_list = FALSE, sv, ids, st, ...)

\method{sur_mean}{character}(x, ...)

surdes_mean(x, ids, sv)

repdes_mean(x, ids, sv)

admin_mean(x)
}
\arguments{
\item{x}{vector. Variable to compute a mean}

\item{...}{unused}

\item{na.rm}{logical. Determines whether NA values will be omitted from the analysis}

\item{as_list}{logical. Determines whether the result should returned as a list-- mostly for factors}

\item{sv}{data.table. Survey vars data.table}

\item{ids}{numeric. vector of indices to operate on}

\item{st}{character. survey type}
}
\description{
Calculate the survey weighted mean of an object

Calculate the mean from a normal survey

calculate the mean for a replicate survey

mean for a admin dataset
}
