% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_uncertainty.R
\name{sur_se}
\alias{sur_se}
\title{Calculate the standard error}
\usage{
sur_se(v, input_type = "var", svyrep_attributes, sv, ids, st)
}
\arguments{
\item{v}{either a vector of data or a vcov matrix}

\item{input_type}{Character. One of 'var' or 'x'. Flags what v represents (x for vector of data) or var for an already computed vcov matrix (usually from sur_var)}

\item{svyrep_attributes}{list. Containing arguments passed to \code{sur_var} when computing the variance from a replicate weight design}
}
\description{
Calculate the standard error
}
